+++ core/src/Decimal.kt
@@ -46,9 +46,9 @@ class Decimal internal constructor(private val c: Long): Comparable<Decimal>, Nu
   fun abs() = if (c < 0) Decimal(-c) else this
   infix fun percent(p: Decimal) = times(p) / CENTS
 
-  override fun equals(o: Any?) = c == (o as? Decimal)?.c
+  override fun equals(other: Any?) = c == (other as? Decimal)?.c
   override fun hashCode() = c.hashCode()
-  override fun compareTo(o: Decimal) = c.compareTo(o.c)
+  override fun compareTo(other: Decimal) = c.compareTo(other.c)
 
   override fun toString(): String {
     val sb = StringBuilder((c / CENTS).toString())
@@ -64,7 +64,6 @@ class Decimal internal constructor(private val c: Long): Comparable<Decimal>, Nu
   override fun toInt() = toLong().toInt()
   override fun toShort() = toLong().toShort()
   override fun toByte() = toLong().toByte()
-  override fun toChar() = toInt().toChar()
 }
 
 val Int.d get() = Decimal(multiplyExact(CENTS, this))