+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * server: ErrorHandler will now omit the default message thrown by first() function ("List/Collection is empty.") and generate a standard 404 NotFound response
 * server: ThrowableHandler now has HttpExchange as receiver for cleaner code, as it is not used in most handlers anyway
 * server: ThrowableHandler returning null will now proceed with next handler, eventually producing error 500 if not handled
+* openapi: new module that can generate OpenAPI json spec file for all routes in a context
 
 # 1.5.6
 * jdbc: DB_READONLY=true env var can be used to make the whole app read-only (e.g. when migrating the DB, instead of Heroku maintenance mode)